+++ api/v1alpha2/workspace_types.go
@@ -329,6 +329,22 @@ type SSHKey struct {
 // +kubebuilder:validation:Pattern:="^[A-Za-z0-9][A-Za-z0-9:_-]*$"
 type Tag string
 
+// DeletionPolicy defines the strategy the Kubernetes operator uses when you delete a resource, either manually or by a system event.
+//
+// You must use one of the following values:
+// - `retain`: When you delete the custom resource, the operator does not delete the workspace.
+// - `soft`: Attempts to delete the associated workspace only if it does not contain any managed resources.
+// - `destroy`: Executes a destroy operation to remove all resources managed by the associated workspace. Once the destruction of these resources is successful, the operator deletes the workspace, and then deletes the custom resource.
+// - `force`: Forcefully and immediately deletes the workspace and the custom resource.
+type DeletionPolicy string
+
+const (
+	DeletionPolicyRetain  DeletionPolicy = "retain"
+	DeletionPolicySoft    DeletionPolicy = "soft"
+	DeletionPolicyDestroy DeletionPolicy = "destroy"
+	DeletionPolicyForce   DeletionPolicy = "force"
+)
+
 // NotificationTrigger represents the different TFC notifications that can be sent as a run's progress transitions between different states.
 // This must be aligned with go-tfe type `NotificationTriggerType`.
 // Must be one of the following values: `run:applying`, `assessment:check_failure`, `run:completed`, `run:created`, `assessment:drifted`, `run:errored`, `assessment:failed`, `run:needs_attention`, `run:planning`.
@@ -561,6 +577,17 @@ type WorkspaceSpec struct {
 	//
 	//+optional
 	Project *WorkspaceProject `json:"project,omitempty"`
+	// The Deletion Policy specifies the behavior of the custom resource and its associated workspace when the custom resource is deleted.
+	// - `retain`: When you delete the custom resource, the operator does not delete the workspace.
+	// - `soft`: Attempts to delete the associated workspace only if it does not contain any managed resources.
+	// - `destroy`: Executes a destroy operation to remove all resources managed by the associated workspace. Once the destruction of these resources is successful, the operator deletes the workspace, and then deletes the custom resource.
+	// - `force`: Forcefully and immediately deletes the workspace and the custom resource.
+	// Default: `retain`.
+	//
+	//+kubebuilder:validation:Enum:=retain;soft;destroy;force
+	//+kubebuilder:default=retain
+	//+optional
+	DeletionPolicy DeletionPolicy `json:"deletionPolicy,omitempty"`
 }
 
 type PlanStatus struct {
@@ -641,6 +668,18 @@ type WorkspaceStatus struct {
 	//
 	//+optional
 	Variables []VariableStatus `json:"variables,omitempty"`
+	// Default organization project ID.
+	//
+	//+optional
+	DefaultProjectID string `json:"defaultProjectID,omitempty"`
+	// SSH Key ID.
+	//
+	//+optional
+	SSHKeyID string `json:"sshKeyID,omitempty"`
+	// Workspace Destroy Run ID.
+	//
+	//+optional
+	DestroyRunID string `json:"destroyRunID,omitempty"`
 }
 
 //+kubebuilder:object:root=true